controller: implement the per-host PCIe device mapping
Before this commit, the FaaS Controller uses a hard-coded PCIe device
mapping, and assumes that all hosts have the exact same mapping.
However, it is not true in reality. Different hosts have their own PCIe
mapping.

The commit implements the per-host PCIe device mapping to initialize
each FaaS worker in the cluster. It takes the cluster information
specified in the user-provided configuration file, and passes the
information to the FaaS Controller class.