cloudlab: sync for debugging experiments in cloudlab
This commit is a start point for working on the cloudlab experiments.
The kubectl module receives the cluster information (the FaaS master
node IP) when the controller is initialized. Therefore, all containers
will set up RPC connections with the FaaS master with the correct IP.

Further, this commit updates the default container choices for both the
NF runtime and the Cooperative Scheduler.