controller: fix a bug of attaching SGroups to unmanaged cores
Prior to this commit, FaaS Controller has a mismatched CPU core set
compared to the Cooperative Scheduler's implementation. In this case,
FaaS Controller may detach a SGroup from core 1, and attach it to an
un-managed core, such as core 18. Then, the SGroup is not managed by any
cores, and thus leads to errors.

This commit fixes the bug by synchronizing the number of cores in FaaS
Controller and in Cooperative Scheduler. Besides, it also refactors the
Core's implementation to display active and idle SGroups separately.
This provides more information for debugging the system.