of_script: add debug logs for collecting per-flow pkt statistics
This commit adds a debug functionality for collecting per-flow pkt
statistics. This is to understand the number of packet arrivals from a
flow before a corresponding flow rule is installed into ToR.

For an one-minute measurment, here are some numbers:
Min: 4, 50-pct: 109, 95-pct: 132, max: 209.
These packets were lost because they do not hit a flow rule, and were
forwarded to the Ryu controller. The Ryu controller is designed to
ingore these subsequent packets.