of_script: configure the idle timeout for all flow rules
Prior to this commit, all flow rules are permanent and not evicted from
the switch. However, according to our experiments, the number of flow
rules in a ToR switch is limited (about 1600 flows). Thus, it is
necessary to set an idle timeout for all flow rules in the ToR switch.

This commit sets an idle timeout of 30 seconds for all flow rules. To
minimize the impact of competing slow-path operations, Ryu controller
now sends out a monitoring OF request once per 60 seconds.