of_script: add debug logs for lost packet rate
This commit adds debug logs for lost packet rate. The Ryu controller
receives subsequent packets from a flow before it installs a rule to
ToR. There packets are simply ignored by the Ryu controller. So, this
commit counts the number of lost packets and calculates the rate of
packet losses.

This information is printed along with other packet-level switch
information.